Move constants from tq to adminserver and common.
